reorganize routers under modules
