+++ src/api/server.py
@@ -3,16 +3,11 @@
 from os import getenv
 
 import sentry_sdk
-from fastapi import APIRouter, FastAPI
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from imsosorry import uwuify
-
-# pylint: disable-next=no-name-in-module
-from pydantic import BaseModel
 
 from . import __version__
-from .models import ServerMetadata
-from .modules.generators import router_generators
+from .modules import routers
 
 release_prefix = getenv("API_SENTRY_RELEASE_PREFIX", "api")
 git_sha = getenv("GIT_SHA", "development")
@@ -24,12 +19,6 @@ sentry_sdk.init(
     profiles_sample_rate=1.0,
     release=f"{release_prefix}@{git_sha}",
 )
-
-
-class TextModel(BaseModel):
-    """Generic model for accepting arbitrary plain-text input"""
-
-    text: str
 
 
 app = FastAPI(
@@ -45,28 +34,6 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-router_root = APIRouter()
 
-
-@router_root.get("/")
-async def metadata() -> ServerMetadata:
-    """Get server metadata"""  
-    return ServerMetadata(
-        version=__version__,
-        server_commit=getenv("GIT_SHA", "development"),
-    )
-
-
-app.include_router(router_root)
-
-router_fun = APIRouter(prefix="/fun", tags=["fun"])
-
-
-@router_fun.post("/uwuify/")
-async def uwuify_route(text: TextModel):
-    """Convert text to UwU meme style"""
-    return {"text": uwuify(text.text)}
-
-
-app.include_router(router_fun)
-app.include_router(router_generators)
+for router in routers:
+    app.include_router(router)
